fix: apply PHP 8.1 change to the strpos method
The strpos method allows the first argument as null and returns null on
previous PHP versions, but raises an exception on PHP 8.1 used on
Magento 2.4.4.

Due to this issue, the category products view shows no products.

This change handles this null value before calling the strpos method to
fix this issue.